personalSite/app/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
codecov==2.0.22
coverage==4.5.4
Django==3.0.2
django-ckeditor==6.1.0
django-js-asset==1.2.2
djangorestframework==3.9.4
flake8==3.6.0
gunicorn==20.0.4
idna==3.2
mccabe==0.6.1
Pillow==7.0.0
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
